Renamed routes to be consistant.
