+++ controllers/api/index.js
@@ -1,10 +1,10 @@
 const router = require('express').Router();
 
 const userRoutes = require('./user-routes');
-const mealRoutes = require('./Meal');
+const mealRoutes = require('./meal-routes');
 
 //js routes
-router.use('/user-routes', userRoutes);
-router.use('/Meal', mealRoutes);
+router.use('/user', userRoutes);
+router.use('/meal', mealRoutes);
 
 module.exports = router;
