+++ models/Meal.js
@@ -58,7 +58,7 @@ Meal.init(
         model: "user",
           key: "id"
       }
-    }         
+    }
   },
   {
     sequelize,
