+++ seeds/userData.json
@@ -1,27 +1,27 @@
 [
       {
-        "username": "Dana Ritchie",
-        "email": "Ned.Homenick38@hotmail.com",
-        "password": "N9879gAxSXBHSFd"
+        "username": "Tamara Quitzon",
+        "email": "Sibyl_Ullrich@hotmail.com",
+        "password": "VjRKoEcgGkRLu5j"
       },
       {
-        "username": "Lynda Torp",
-        "email": "Rhiannon51@hotmail.com",
-        "password": "AAqgDwDki8sB2yO"
+        "username": "Dana Halvorson",
+        "email": "Roman.Ortiz78@gmail.com",
+        "password": "60Xbp99AjZYmWIA"
       },
       {
-        "username": "Seth Smitham",
-        "email": "Rosetta_Lind@gmail.com",
-        "password": "Wwl4Am8JBLy7s9x"
+        "username": "Lee Langosh",
+        "email": "Marley_Donnelly@yahoo.com",
+        "password": "X9hxK58iVW4JOYI"
       },
       {
-        "username": "Cody Kunde",
-        "email": "Amelie4@hotmail.com",
-        "password": "x4OcAaXdlVsT0Y6"
+        "username": "Oscar Mann MD",
+        "email": "Garrison_Aufderhar@gmail.com",
+        "password": "tS9gK4nR9Jpgbfo"
       },
       {
-        "username": "Melanie Crist",
-        "email": "Chris.Kuphal@yahoo.com",
-        "password": "HvWcUj0Vf1MllCU"
+        "username": "Geraldine Sauer",
+        "email": "Addie94@yahoo.com",
+        "password": "LWojgMZe_cHSYPh"
       }
     ]
